Change saved data, use only changed attributes
There is no need to keep all the original attributes of a changed model. Only those that have a change are relevant, e.g. so developers can show "old" and "new" values for attributes that have been changed, instead of all.
Also, changed observer listener from "updated" to "updating", so that dirty fields can be obtained before they are updated.