Change provide, publish migrations instead of loading them directly
Loading the migrations is a poor choice for this package. Developers should be allowed to modify the audit_trail table as much as they wish, including adding more columns. Furthermore, Renaming the migration's date time can also be required, to change the migration's installation order (depending on circumstance). This was previously not possible.